fix: wishlist_recommendwork 요소 수정

+++ wishlist/models.py
@@ -62,7 +62,7 @@ class Contents(models.Model):
 class RecommendedWork(models.Model):
     recommended_id = models.AutoField(primary_key=True)
     account_user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
-    content_id = models.IntegerField()
+    content_id = models.BigIntegerField()
     recommended_model = models.CharField(max_length=50)
     recommended_date = models.DateTimeField(auto_now_add=True, null=True)
     feedback = models.IntegerField(null=True, blank=True)
